worker/pyproject.toml
[project]
name = "markets-o11y-fetcher"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "yfinance>=0.2.40",
    "psycopg2-binary>=2.9",
    "pyyaml>=6.0",
    "schedule>=1.2",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["app"]
